+++ NEWS.md
@@ -1,3 +1,7 @@
+# plantRdata 0.0.9
+
+* Updating all backbones
+
 # plantRdata 0.0.8
 
 * Fixing the BFO backbone regarding the presence of quadrinomials in the datatset (subforms and subvarieties)